+++ e2e-tests/production-runtime/src/pages/client-only-paths.js
@@ -1,4 +1,4 @@
-import React from 'react'
+import * as React from 'react'
 import { Router } from '@reach/router'
 import { Link } from 'gatsby'