+++ packages/gatsby/src/commands/develop.ts
@@ -31,13 +31,13 @@ import * as WorkerPool from "../utils/worker/pool"
 import http from "http"
 import https from "https"
 
-import bootstrapSchemaHotReloader from "../bootstrap/schema-hot-reloader"
+import { bootstrapSchemaHotReloader } from "../bootstrap/schema-hot-reloader"
 import bootstrapPageHotReloader from "../bootstrap/page-hot-reloader"
 import { developStatic } from "./develop-static"
 import withResolverContext from "../schema/context"
 import sourceNodes from "../utils/source-nodes"
-import createSchemaCustomization from "../utils/create-schema-customization"
-import websocketManager from "../utils/websocket-manager"
+import { createSchemaCustomization } from "../utils/create-schema-customization"
+import { websocketManager } from "../utils/websocket-manager"
 import getSslCert from "../utils/get-ssl-cert"
 import { slash } from "gatsby-core-utils"
 import { initTracer } from "../utils/tracer"
@@ -198,7 +198,7 @@ async function startServer(program: IProgram): Promise<IServer> {
 
   /**
    * Refresh external data sources.
-   * This behavior is disabled by default, but the ENABLE_REFRESH_ENDPOINT env var enables it
+   * This behavior is disabled by default, but the ENABLE_GATSBY_REFRESH_ENDPOINT env var enables it
    * If no GATSBY_REFRESH_TOKEN env var is available, then no Authorization header is required
    **/
   const REFRESH_ENDPOINT = `/__refresh`
@@ -320,8 +320,7 @@ async function startServer(program: IProgram): Promise<IServer> {
     ? https.createServer(program.ssl, app)
     : new http.Server(app)
 
-  websocketManager.init({ server, directory: program.directory })
-  const socket = websocketManager.getSocket()
+  const socket = websocketManager.init({ server, directory: program.directory })
 
   const listener = server.listen(program.port, program.host)